Rename `retry_budget` module
