+++ ydb/src/client_topic/topicreader/reader_options.rs
@@ -1,6 +1,6 @@
 use crate::client_topic::compression::CompressionDecoder;
 use crate::client_topic::topicreader::reader::TopicSelectors;
-use crate::retry_budget::{ArcRetrySettings, RetrySettings};
+use crate::retry_settings::{ArcRetrySettings, RetrySettings};
 use std::sync::Arc;
 
 #[derive(bon::Builder, Clone)]
